refactor(ai): split explainer responsibilities
Extract AI explanation data models, prompts, OpenRouter client calls, text-product handling, prompting, and validation helpers into focused modules while preserving the existing ai_explainer import surface.

Tested: uv run ruff check src tests scripts; uv run pyright; uv run pytest -q -n 0 --tb=short

Confidence: high
