+++ src/accessiweather/ai_explainer.py
@@ -7,287 +7,66 @@ using OpenRouter's unified API gateway for AI models.
 
 from __future__ import annotations
 
-import hashlib
 import logging
-import re
-from dataclasses import dataclass
 from datetime import datetime
-from enum import Enum
-from typing import TYPE_CHECKING, Any, Literal
+from typing import TYPE_CHECKING, Any
+
+from .ai_explainer_models import (
+    AIExplainerError,
+    EmptyResponseError,
+    ExplanationResult,
+    ExplanationStyle,
+    InsufficientCreditsError,
+    InvalidAPIKeyError,
+    InvalidModelError,
+    NetworkError,
+    RateLimitError,
+    RequestTimeoutError,
+    TextProductType,
+    WeatherContext,
+)
+from .ai_explainer_openrouter import (
+    DEFAULT_FREE_MODEL,
+    DEFAULT_FREE_ROUTER,
+    get_available_free_models,
+)
+from .ai_explainer_openrouter_client import AIExplainerOpenRouterMixin
+from .ai_explainer_prompting import AIExplainerPromptMixin
+from .ai_explainer_prompts import SYSTEM_PROMPTS as _SYSTEM_PROMPTS
+from .ai_explainer_text_products import AIExplainerTextProductMixin
+from .ai_explainer_validation import AIExplainerValidationMixin
 
 if TYPE_CHECKING:
     from .cache import Cache
 
-# Product-type literal used by explain_text_product. Keep in sync with
-# _SYSTEM_PROMPTS keys.
-TextProductType = Literal["AFD", "HWO", "SPS"]
-
 logger = logging.getLogger(__name__)
 
-# OpenRouter API configuration
-OPENROUTER_BASE_URL = "https://openrouter.ai/api/v1"
-OPENROUTER_MODELS_URL = "https://openrouter.ai/api/v1/models"
-# Use current working free models from OpenRouter (updated Dec 2025)
-DEFAULT_FREE_MODEL = "openrouter/free"
-DEFAULT_FREE_ROUTER = "openrouter/free"
-FALLBACK_FREE_MODEL = "meta-llama/llama-3.3-70b-instruct:free"
-DEFAULT_PAID_MODEL = "openrouter/auto"
-# Static fallback models only used if dynamic fetch fails
-STATIC_FALLBACK_MODELS = [
-    "google/gemma-3-27b-it:free",
-    "qwen/qwen3-4b:free",
+__all__ = [
+    "AIExplainer",
+    "AIExplainerError",
+    "EmptyResponseError",
+    "ExplanationResult",
+    "ExplanationStyle",
+    "InsufficientCreditsError",
+    "InvalidAPIKeyError",
+    "InvalidModelError",
+    "NetworkError",
+    "RateLimitError",
+    "RequestTimeoutError",
+    "TextProductType",
+    "WeatherContext",
+    "_SYSTEM_PROMPTS",
+    "get_available_free_models",
+    "has_valid_api_key",
 ]
 
-# Cache for dynamically fetched free models
-_free_models_cache: list[str] | None = None
-_free_models_cache_time: float = 0
-_FREE_MODELS_CACHE_TTL = 3600  # 1 hour cache
 
-
-def get_available_free_models(exclude_model: str | None = None) -> list[str]:
-    """
-    Fetch available free models from OpenRouter API.
-
-    Results are cached for 1 hour to avoid excessive API calls.
-    Falls back to static list if API is unavailable.
-
-    Args:
-        exclude_model: Model ID to exclude from results (e.g., the primary model)
-
-    Returns:
-        List of free model IDs, up to 3 models (OpenRouter's fallback limit)
-
-    """
-    import time
-
-    global _free_models_cache, _free_models_cache_time
-
-    # Return cached results if still valid
-    if (
-        _free_models_cache is not None
-        and (time.time() - _free_models_cache_time) < _FREE_MODELS_CACHE_TTL
-    ):
-        models = [m for m in _free_models_cache if m != exclude_model]
-        return models[:3]
-
-    try:
-        import httpx
-
-        response = httpx.get(OPENROUTER_MODELS_URL, timeout=10.0)
-        response.raise_for_status()
-        data = response.json()
-
-        # Filter for free models and sort by context length (prefer larger context)
-        free_models = []
-        for model in data.get("data", []):
-            model_id = model.get("id", "")
-            if model_id.endswith(":free"):
-                context_length = model.get("context_length", 0)
-                free_models.append((model_id, context_length))
-
-        # Sort by context length descending, take top models
-        free_models.sort(key=lambda x: x[1], reverse=True)
-        _free_models_cache = [m[0] for m in free_models]
-        _free_models_cache_time = time.time()
-
-        logger.info(f"Fetched {len(_free_models_cache)} free models from OpenRouter")
-
-        models = [m for m in _free_models_cache if m != exclude_model]
-        return models[:3]
-
-    except Exception as e:
-        logger.warning(f"Failed to fetch free models from OpenRouter: {e}, using static fallbacks")
-        models = [m for m in STATIC_FALLBACK_MODELS if m != exclude_model]
-        return models[:3]
-
-
-# System prompts for generic NWS text-product explanations (AFD / HWO / SPS).
-#
-# IMPORTANT: The AFD prompt below is the historical explain_afd system prompt,
-# preserved byte-for-byte. User-customized prompt overrides in the wild are
-# calibrated against this exact wording. Do not reformat or rephrase without
-# treating it as a behavior change.
-_SYSTEM_PROMPTS: dict[str, str] = {
-    "AFD": (
-        "You are a helpful weather assistant that explains National Weather Service "
-        "Area Forecast Discussions (AFDs) in plain, accessible language. AFDs contain "
-        "technical meteorological terminology that most people don't understand. "
-        "Your job is to translate this into clear, everyday language that anyone can "
-        "understand. Focus on:\n"
-        "- What weather to expect and when\n"
-        "- Any significant weather events or changes\n"
-        "- How confident forecasters are in their predictions\n"
-        "- What this means for daily activities\n\n"
-        "Avoid using technical jargon. If you must use a technical term, explain it.\n\n"
-        "IMPORTANT: Do NOT start with a preamble like 'Here is a summary...' or "
-        "'This forecast discussion explains...'. Do NOT repeat the location name. "
-        "Jump straight into explaining the weather. The user already knows what they asked for.\n\n"
-        "IMPORTANT: Respond in plain text only. Do NOT use markdown formatting such as "
-        "bold (**text**), italic (*text*), headers (#), bullet points, or any other "
-        "markdown syntax. Use simple paragraph text that can be read directly."
-    ),
-    "HWO": (
-        "You are a helpful weather assistant that explains National Weather Service "
-        "Hazardous Weather Outlooks (HWOs) in plain, accessible language. An HWO "
-        "describes the 7-day hazard horizon and probabilistic outlook for severe "
-        "weather, flooding, winter weather, marine hazards, and other significant "
-        "events. Your job is to translate it into clear, everyday language that "
-        "anyone can understand. Focus on:\n"
-        "- Which hazards are on the horizon and when they may arrive\n"
-        "- How likely or confident the outlook is (e.g., slight, moderate, high chance)\n"
-        "- What areas or activities are most affected\n"
-        "- What this means for daily plans over the coming week\n\n"
-        "Avoid using technical jargon. If you must use a technical term, explain it.\n\n"
-        "IMPORTANT: Do NOT start with a preamble like 'Here is a summary...' or "
-        "'This outlook explains...'. Do NOT repeat the location name. Jump straight "
-        "into explaining the hazards. The user already knows what they asked for.\n\n"
-        "IMPORTANT: Respond in plain text only. Do NOT use markdown formatting such as "
-        "bold (**text**), italic (*text*), headers (#), bullet points, or any other "
-        "markdown syntax. Use simple paragraph text that can be read directly."
-    ),
-    "SPS": (
-        "You are a helpful weather assistant that explains National Weather Service "
-        "Special Weather Statements (SPSs) in plain, accessible language. An SPS is "
-        "a short advisory about significant but sub-warning weather — strong "
-        "thunderstorms, dense fog, brief heavy rain, localized wind — that people "
-        "should know about but that doesn't rise to a formal warning. Your job is "
-        "to summarize it in clear, everyday language. Focus on:\n"
-        "- What the advisory is about and how long it lasts\n"
-        "- Who and where it affects\n"
-        "- What people in the affected area should do\n"
-        "- How serious the threat is compared to a warning\n\n"
-        "Avoid using technical jargon. If you must use a technical term, explain it.\n\n"
-        "IMPORTANT: Do NOT start with a preamble like 'Here is a summary...' or "
-        "'This statement explains...'. Do NOT repeat the location name. Jump "
-        "straight into explaining the advisory. The user already knows what they asked for.\n\n"
-        "IMPORTANT: Respond in plain text only. Do NOT use markdown formatting such as "
-        "bold (**text**), italic (*text*), headers (#), bullet points, or any other "
-        "markdown syntax. Use simple paragraph text that can be read directly."
-    ),
-}
-
-
-class AIExplainerError(Exception):
-    """Base exception for AI explainer errors."""
-
-
-class InsufficientCreditsError(AIExplainerError):
-    """Raised when OpenRouter account has no funds."""
-
-
-class RateLimitError(AIExplainerError):
-    """Raised when rate limits are exceeded."""
-
-
-class InvalidAPIKeyError(AIExplainerError):
-    """Raised when API key is invalid or malformed."""
-
-
-class InvalidModelError(AIExplainerError):
-    """Raised when the specified model ID does not exist."""
-
-
-class NetworkError(AIExplainerError):
-    """Raised when network connectivity issues occur."""
-
-
-class RequestTimeoutError(AIExplainerError):
-    """Raised when API request times out."""
-
-
-class EmptyResponseError(AIExplainerError):
-    """Raised when all models return empty or insufficient responses."""
-
-
-class ExplanationStyle(Enum):
-    """Available explanation styles."""
-
-    BRIEF = "brief"  # 1-2 sentences
-    STANDARD = "standard"  # 3-4 sentences (default)
-    DETAILED = "detailed"  # Full paragraph with context
-
-
-@dataclass
-class ExplanationResult:
-    """Result of an explanation generation."""
-
-    text: str
-    model_used: str
-    token_count: int
-    estimated_cost: float
-    cached: bool
-    timestamp: datetime
-
-
-@dataclass
-class WeatherContext:
-    """Structured weather data for AI explanation."""
-
-    location: str
-    timestamp: datetime
-    temperature: float | None
-    temperature_unit: str
-    conditions: str | None
-    humidity: int | None
-    wind_speed: float | None
-    wind_direction: str | None
-    visibility: float | None
-    pressure: float | None
-    alerts: list[dict[str, Any]]
-    forecast_summary: str | None = None
-    local_time: str | None = None
-    utc_time: str | None = None
-    timezone: str | None = None
-    time_of_day: str | None = None
-
-    def to_prompt_text(self) -> str:
-        """Convert to natural language for AI prompt."""
-        parts = [f"Location: {self.location}"]
-
-        # Add time context so AI knows if it's morning/afternoon/evening/night
-        if self.local_time and self.timezone:
-            parts.append(f"Local Time: {self.local_time} ({self.timezone})")
-        if self.utc_time:
-            parts.append(f"UTC Time: {self.utc_time}")
-        if self.time_of_day:
-            parts.append(f"Time of Day: {self.time_of_day}")
-
-        if self.temperature is not None:
-            parts.append(f"Temperature: {self.temperature}°{self.temperature_unit}")
-
-        if self.conditions:
-            parts.append(f"Conditions: {self.conditions}")
-
-        if self.humidity is not None:
-            parts.append(f"Humidity: {self.humidity}%")
-
-        if self.wind_speed is not None:
-            wind_info = f"Wind: {self.wind_speed} mph"
-            if self.wind_direction:
-                wind_info += f" from {self.wind_direction}"
-            parts.append(wind_info)
-
-        if self.visibility is not None:
-            parts.append(f"Visibility: {self.visibility} miles")
-
-        if self.pressure is not None:
-            parts.append(f"Pressure: {self.pressure} inHg")
-
-        if self.alerts:
-            alert_texts = []
-            for alert in self.alerts:
-                title = alert.get("title", "Weather Alert")
-                severity = alert.get("severity", "Unknown")
-                alert_texts.append(f"- {title} (Severity: {severity})")
-            parts.append("Active Alerts:\n" + "\n".join(alert_texts))
-
-        if self.forecast_summary:
-            parts.append(f"Forecast: {self.forecast_summary}")
-
-        return "\n".join(parts)
-
-
-class AIExplainer:
+class AIExplainer(
+    AIExplainerPromptMixin,
+    AIExplainerOpenRouterMixin,
+    AIExplainerTextProductMixin,
+    AIExplainerValidationMixin,
+):
     """Generates natural language weather explanations using OpenRouter."""
 
     def __init__(
@@ -317,55 +96,6 @@ class AIExplainer:
         self._session_token_count = 0
         self._client = None
 
-    # Maximum allowed length for custom prompts to prevent abuse
-    _MAX_PROMPT_LENGTH = 2000
-
-    @staticmethod
-    def _sanitize_prompt(prompt: str | None) -> str | None:
-        """
-        Sanitize a custom prompt to prevent prompt injection attacks.
-
-        Enforces length limits and strips potentially dangerous patterns
-        like attempts to override system behavior or extract data.
-
-        Args:
-            prompt: The custom prompt text to sanitize
-
-        Returns:
-            Sanitized prompt or None if input was None/empty
-
-        """
-        if not prompt or not prompt.strip():
-            return None
-
-        sanitized = prompt.strip()
-
-        # Enforce length limit
-        if len(sanitized) > AIExplainer._MAX_PROMPT_LENGTH:
-            logger.warning(
-                f"Custom prompt truncated from {len(sanitized)} to "
-                f"{AIExplainer._MAX_PROMPT_LENGTH} characters"
-            )
-            sanitized = sanitized[: AIExplainer._MAX_PROMPT_LENGTH]
-
-        # Strip common prompt injection patterns
-        injection_patterns = [
-            r"(?i)ignore\s+(all\s+)?(previous|above|prior)\s+(instructions?|prompts?|rules?)",
-            r"(?i)you\s+are\s+now\s+(?:a|an|in)\s+(?:new|different|unrestricted)",
-            r"(?i)disregard\s+(all\s+)?(previous|above|prior|your)\s+(instructions?|programming|rules?)",
-            r"(?i)system\s*:\s*",
-            r"(?i)\n\s*system\s*:",
-        ]
-
-        import re as _re
-
-        for pattern in injection_patterns:
-            if _re.search(pattern, sanitized):
-                logger.warning(f"Prompt injection pattern detected and stripped: {pattern}")
-                sanitized = _re.sub(pattern, "[filtered]", sanitized)
-
-        return sanitized if sanitized.strip() else None
-
     def get_effective_model(self) -> str:
         """
         Get the effective model based on API key and preference.
@@ -380,233 +110,6 @@ class AIExplainer:
 
         # With API key, use configured preference (fall back to default if None)
         return self.model if self.model else DEFAULT_FREE_MODEL
-
-    def _get_client(self):
-        """Get or create OpenAI client configured for OpenRouter."""
-        if self._client is None:
-            # OpenRouter requires an API key for ALL requests, including free models
-            # The :free suffix only means no charges, not no authentication
-            if not self.api_key:
-                raise AIExplainerError(
-                    "OpenRouter API key required. Get a free key at openrouter.ai/keys - "
-                    "free models won't charge your account."
-                )
-
-            try:
-                from openai import OpenAI
-
-                self._client = OpenAI(
-                    base_url=OPENROUTER_BASE_URL,
-                    api_key=self.api_key,
-                    timeout=30.0,  # 30 second timeout to prevent hanging
-                )
-            except ImportError as e:
-                logger.error("OpenAI package not installed")
-                raise AIExplainerError(
-                    "AI explanation feature requires the openai package. "
-                    "Please install it with: pip install openai"
-                ) from e
-        return self._client
-
-    def get_effective_system_prompt(self, style: ExplanationStyle) -> str:
-        """
-        Get the system prompt to use, preferring custom over default.
-
-        Args:
-            style: Explanation style for default prompt
-
-        Returns:
-            Custom system prompt if configured, otherwise default prompt
-
-        """
-        if self.custom_system_prompt:
-            return self.custom_system_prompt
-        return self._build_default_system_prompt(style)
-
-    def _build_default_system_prompt(self, style: ExplanationStyle) -> str:
-        """Build the default system prompt based on explanation style."""
-        base_prompt = self.get_default_system_prompt()
-
-        style_instructions = {
-            ExplanationStyle.BRIEF: "Keep your response to 1-2 sentences.",
-            ExplanationStyle.STANDARD: "Provide a 3-4 sentence explanation.",
-            ExplanationStyle.DETAILED: (
-                "Provide a comprehensive paragraph with context about "
-                "how the weather might affect daily activities."
-            ),
-        }
-
-        return f"{base_prompt}\n\n{style_instructions.get(style, style_instructions[ExplanationStyle.STANDARD])}"
-
-    @staticmethod
-    def get_default_system_prompt() -> str:
-        """
-        Return the default system prompt text for display in UI.
-
-        Returns:
-            The default system prompt string
-
-        """
-        return (
-            "You are a helpful weather assistant that explains weather conditions "
-            "in plain, accessible language. Your explanations should be easy to "
-            "understand for screen reader users and people who prefer audio descriptions. "
-            "Avoid using visual-only descriptions. Focus on how the weather will feel "
-            "and what activities it's suitable for.\n\n"
-            "IMPORTANT: Do NOT repeat the location name, date, time, or timezone in your response. "
-            "The user already sees this information. Jump straight into describing the weather "
-            "conditions and what to expect.\n\n"
-            "IMPORTANT: Respond in plain text only. Do NOT use markdown formatting such as "
-            "bold (**text**), italic (*text*), headers (#), bullet points, or any other "
-            "markdown syntax. Use simple paragraph text."
-        )
-
-    def get_prompt_preview(
-        self,
-        style: ExplanationStyle = ExplanationStyle.STANDARD,
-    ) -> dict[str, str]:
-        """
-        Generate a preview of the prompts that will be sent to the AI.
-
-        Args:
-            style: Explanation style to use for the preview
-
-        Returns:
-            Dict with 'system_prompt' and 'user_prompt' keys
-
-        """
-        sample_weather = {
-            "temperature": 72,
-            "temperature_unit": "F",
-            "conditions": "Partly Cloudy",
-            "humidity": 65,
-            "wind_speed": 8,
-            "wind_direction": "NW",
-            "visibility": 10,
-        }
-        return {
-            "system_prompt": self.get_effective_system_prompt(style),
-            "user_prompt": self._build_prompt(sample_weather, "Sample Location", style),
-        }
-
-    def _build_prompt(
-        self,
-        weather_data: dict[str, Any],
-        location_name: str,
-        style: ExplanationStyle,
-    ) -> str:
-        """Construct prompt from weather data."""
-        # Extract weather context from data
-        context = WeatherContext(
-            location=location_name,
-            timestamp=datetime.now(),
-            temperature=weather_data.get("temperature"),
-            temperature_unit=weather_data.get("temperature_unit", "F"),
-            conditions=weather_data.get("conditions"),
-            humidity=weather_data.get("humidity"),
-            wind_speed=weather_data.get("wind_speed"),
-            wind_direction=weather_data.get("wind_direction"),
-            visibility=weather_data.get("visibility"),
-            pressure=weather_data.get("pressure"),
-            alerts=weather_data.get("alerts", []),
-            forecast_summary=weather_data.get("forecast_summary"),
-            local_time=weather_data.get("local_time"),
-            utc_time=weather_data.get("utc_time"),
-            timezone=weather_data.get("timezone"),
-            time_of_day=weather_data.get("time_of_day"),
-        )
-
-        prompt_parts = [
-            "Please explain the following weather conditions:\n",
-            context.to_prompt_text(),
-        ]
-
-        # Add forecast periods if available
-        forecast_periods = weather_data.get("forecast_periods", [])
-        if forecast_periods:
-            prompt_parts.append("\n\nUpcoming Forecast:")
-            for period in forecast_periods:
-                period_text = f"\n- {period.get('name', 'Unknown')}: "
-                period_text += (
-                    f"{period.get('temperature', 'N/A')}°{period.get('temperature_unit', 'F')}"
-                )
-                if period.get("short_forecast"):
-                    period_text += f", {period['short_forecast']}"
-                if period.get("wind_speed"):
-                    period_text += f" (Wind: {period['wind_speed']}"
-                    if period.get("wind_direction"):
-                        period_text += f" {period['wind_direction']}"
-                    period_text += ")"
-                prompt_parts.append(period_text)
-
-        prompt_parts.append(
-            "\n\nProvide a natural language explanation of the current conditions "
-            "and what to expect over the coming days for someone planning their activities."
-        )
-
-        # Add custom instructions if configured
-        if self.custom_instructions and self.custom_instructions.strip():
-            prompt_parts.append(f"\n\nAdditional Instructions: {self.custom_instructions}")
-
-        return "".join(prompt_parts)
-
-    def _format_response(self, response_text: str, preserve_markdown: bool) -> str:
-        """
-        Format AI response based on HTML rendering setting.
-
-        Args:
-            response_text: Raw response from AI
-            preserve_markdown: If True, keep markdown; if False, strip it
-
-        Returns:
-            Formatted response text
-
-        """
-        if preserve_markdown:
-            return response_text
-
-        # Strip markdown formatting for plain text output
-        text = response_text
-
-        # Remove bold/italic markers
-        text = re.sub(r"\*\*(.+?)\*\*", r"\1", text)  # **bold**
-        text = re.sub(r"\*(.+?)\*", r"\1", text)  # *italic*
-        text = re.sub(r"__(.+?)__", r"\1", text)  # __bold__
-        text = re.sub(r"_(.+?)_", r"\1", text)  # _italic_
-
-        # Remove headers
-        text = re.sub(r"^#{1,6}\s+", "", text, flags=re.MULTILINE)
-
-        # Remove code blocks
-        text = re.sub(r"```[\s\S]*?```", "", text)
-        text = re.sub(r"`(.+?)`", r"\1", text)
-
-        # Remove links but keep text
-        text = re.sub(r"\[(.+?)\]\(.+?\)", r"\1", text)
-
-        # Remove bullet points
-        text = re.sub(r"^\s*[-*+]\s+", "", text, flags=re.MULTILINE)
-
-        # Clean up extra whitespace
-        text = re.sub(r"\n{3,}", "\n\n", text)
-
-        return text.strip()
-
-    def _generate_cache_key(self, weather_data: dict[str, Any], location_name: str) -> str:
-        """Generate a cache key for the explanation."""
-        # Create a key based on significant weather values
-        key_parts = [
-            f"loc:{location_name}",
-            f"temp:{weather_data.get('temperature')}",
-            f"cond:{weather_data.get('conditions')}",
-            f"model:{self.get_effective_model()}",
-        ]
-        return "ai_explanation:" + ":".join(key_parts)
-
-    @property
-    def session_token_count(self) -> int:
-        """Get total tokens used in this session."""
-        return self._session_token_count
 
     async def explain_weather(
         self,
@@ -813,640 +316,6 @@ class AIExplainer:
             logger.debug(f"Cached explanation: {cache_key}")
 
         return result
-
-    # Product-type-specific phrasing for the user-message lead-in. The AFD
-    # phrasing is preserved byte-for-byte from the historical explain_afd
-    # implementation to keep downstream prompt templates stable.
-    _PRODUCT_USER_INTRO: dict[str, str] = {
-        "AFD": "Please explain this Area Forecast Discussion for {location} in plain language:",
-        "HWO": "Please explain this Hazardous Weather Outlook for {location} in plain language:",
-        "SPS": "Please explain this Special Weather Statement for {location} in plain language:",
-    }
-
-    # Product-wide style instructions appended to the user prompt. Shared
-    # across AFD/HWO/SPS so customization lives in one place.
-    _PRODUCT_STYLE_INSTRUCTIONS: dict[ExplanationStyle, str] = {
-        ExplanationStyle.BRIEF: "Provide a 2-3 sentence summary of the key points.",
-        ExplanationStyle.STANDARD: "Provide a clear 1-2 paragraph summary.",
-        ExplanationStyle.DETAILED: (
-            "Provide a comprehensive summary covering all major points "
-            "from the discussion, organized by topic."
-        ),
-    }
-
-    def _text_product_cache_key(
-        self,
-        product_type: str,
-        location_name: str,
-        product_text: str,
-        style: ExplanationStyle,
-    ) -> str:
-        """
-        Compute the cache key for explain_text_product results.
-
-        Key shape: ``ai_text_product:<TYPE>:<location>:<sha256>:<style>``.
-        Hashing the product text keeps the key bounded and deterministic for
-        long AFDs/HWOs/SPSs.
-        """
-        text_hash = hashlib.sha256(product_text.encode("utf-8")).hexdigest()
-        return f"ai_text_product:{product_type}:{location_name}:{text_hash}:{style.value}"
-
-    async def explain_text_product(
-        self,
-        product_text: str,
-        product_type: TextProductType,
-        location_name: str,
-        *,
-        style: ExplanationStyle = ExplanationStyle.DETAILED,
-        preserve_markdown: bool = False,
-    ) -> ExplanationResult:
-        """
-        Generate a plain-language explanation of an NWS text product.
-
-        Supports AFD (Area Forecast Discussion), HWO (Hazardous Weather
-        Outlook), and SPS (Special Weather Statement). The system prompt is
-        selected per product type from :data:`_SYSTEM_PROMPTS`. A per-product
-        result cache (300 s TTL) avoids re-invoking the LLM for identical
-        inputs. LLM errors propagate and are not cached — a retry will
-        re-hit the model.
-
-        Args:
-            product_text: Raw product text as issued by NWS.
-            product_type: One of ``"AFD"``, ``"HWO"``, ``"SPS"``.
-            location_name: Human-readable location the product covers.
-            style: Explanation style (brief, standard, detailed).
-            preserve_markdown: Keep markdown in the response when True.
-
-        Returns:
-            :class:`ExplanationResult`. ``cached=True`` if served from the
-            per-product cache.
-
-        """
-        import asyncio
-
-        if product_type not in _SYSTEM_PROMPTS:
-            raise ValueError(
-                f"Unknown product_type {product_type!r}; expected one of {sorted(_SYSTEM_PROMPTS)}"
-            )
-
-        # Cache lookup (custom prompts + instructions are applied per-instance
-        # and already reflected in the request; keying off product_type /
-        # location / text / style is sufficient for a single explainer).
-        cache_key = self._text_product_cache_key(product_type, location_name, product_text, style)
-        if self.cache is not None:
-            cached = self.cache.get(cache_key)
-            if cached:
-                logger.debug(f"Cache hit for text product explanation: {cache_key}")
-                return ExplanationResult(
-                    text=cached["text"],
-                    model_used=cached["model_used"],
-                    token_count=cached["token_count"],
-                    estimated_cost=cached["estimated_cost"],
-                    cached=True,
-                    timestamp=datetime.fromisoformat(cached["timestamp"]),
-                )
-
-        # System prompt: custom overrides the default (preserves existing
-        # explain_afd semantics — replace, not append).
-        system_prompt = self.custom_system_prompt or _SYSTEM_PROMPTS[product_type]
-
-        # User prompt: product-type-aware lead-in + raw product + style hint.
-        intro_template = self._PRODUCT_USER_INTRO[product_type]
-        intro = intro_template.format(location=location_name)
-        style_instruction = self._PRODUCT_STYLE_INSTRUCTIONS.get(
-            style, self._PRODUCT_STYLE_INSTRUCTIONS[ExplanationStyle.DETAILED]
-        )
-        user_prompt = f"{intro}\n\n{product_text}\n\n{style_instruction}"
-
-        # Custom per-user instructions apply identically to all product types.
-        if self.custom_instructions and self.custom_instructions.strip():
-            user_prompt += f"\n\nAdditional Instructions: {self.custom_instructions}"
-
-        # Build list of models to try: primary first, then fallbacks
-        primary_model = self.get_effective_model()
-        models_to_try = [primary_model]
-
-        if primary_model != DEFAULT_FREE_MODEL:
-            models_to_try.append(DEFAULT_FREE_MODEL)
-
-        if ":free" in primary_model or primary_model in (DEFAULT_FREE_MODEL, DEFAULT_FREE_ROUTER):
-            fallback_models = get_available_free_models(exclude_model=primary_model)
-            for fallback in fallback_models:
-                if fallback not in models_to_try:
-                    models_to_try.append(fallback)
-
-        response = None
-        last_error = None
-        for model in models_to_try:
-            try:
-                model_override = model if model != primary_model else None
-                response = await asyncio.to_thread(
-                    self._call_openrouter, system_prompt, user_prompt, model_override
-                )
-
-                content = response["content"]
-                if content and len(content.strip()) >= 20:
-                    logger.info(f"Got valid {product_type} response from model: {model}")
-                    break
-
-                logger.warning(
-                    f"Model {model} returned insufficient {product_type} response "
-                    f"(len={len(content) if content else 0}), trying fallback..."
-                )
-                response = None
-
-            except Exception as e:
-                last_error = e
-                logger.warning(f"Model {model} failed for {product_type}: {e}, trying fallback...")
-                continue
-
-        if response is None:
-            if last_error:
-                logger.error(
-                    f"All models failed for {product_type}. Last error: {last_error}",
-                    exc_info=True,
-                )
-                error_message = str(last_error).lower()
-
-                if "api key" in error_message or "api_key" in error_message:
-                    raise InvalidAPIKeyError(
-                        "OpenRouter API key is required.\n\n"
-                        "Please add your API key in Settings → AI Explanations.\n"
-                        "Get a free key at: openrouter.ai/keys"
-                    ) from last_error
-
-                if (
-                    "404" in error_message
-                    or "not found" in error_message
-                    or "no endpoints found" in error_message
-                    or "does not exist" in error_message
-                ):
-                    raise InvalidModelError(
-                        f"The AI model '{primary_model}' was not found.\n\n"
-                        "It may have been removed or renamed by OpenRouter.\n"
-                        "Please go to Settings → AI Explanations and select a different model."
-                    ) from last_error
-
-                if (
-                    "429" in error_message
-                    or "rate limit" in error_message
-                    or "rate-limited" in error_message
-                ):
-                    raise RateLimitError(
-                        "All AI models are currently rate-limited.\n\n"
-                        "Free models share rate limits with all users.\n\n"
-                        "Options:\n"
-                        "• Wait a few minutes and try again\n"
-                        "• Add credits to your OpenRouter account\n"
-                        "• Switch to a paid model in Settings"
-                    ) from last_error
-
-                if "timed out" in error_message or "timeout" in error_message:
-                    raise RequestTimeoutError(
-                        "Request timed out.\n\n"
-                        "The AI service is taking too long to respond.\n"
-                        "This usually means the servers are busy. Please try again."
-                    ) from last_error
-
-                if (
-                    "502" in error_message
-                    or "503" in error_message
-                    or "network error" in error_message
-                    or "connection refused" in error_message
-                    or "connection reset" in error_message
-                ):
-                    raise NetworkError(
-                        "Network connection error while contacting AI service.\n\n"
-                        "Please check your internet connection and try again."
-                    ) from last_error
-
-                # Re-raise the original error
-                raise last_error
-
-            raise EmptyResponseError(
-                "All AI models returned empty responses.\n\n"
-                "This can happen when models are overloaded.\n"
-                "Please try again in a few minutes."
-            )
-
-        # Success: format + build result
-        raw_content = response["content"]
-        text = self._format_response(raw_content, preserve_markdown)
-        token_count = response["total_tokens"]
-        model_used = response["model"]
-        estimated_cost = self._estimate_cost(model_used, token_count)
-        self._session_token_count += token_count
-
-        result = ExplanationResult(
-            text=text,
-            model_used=model_used,
-            token_count=token_count,
-            estimated_cost=estimated_cost,
-            cached=False,
-            timestamp=datetime.now(),
-        )
-
-        # Cache only successful results — failures must not be memoized.
-        if self.cache is not None:
-            cache_data = {
-                "text": result.text,
-                "model_used": result.model_used,
-                "token_count": result.token_count,
-                "estimated_cost": result.estimated_cost,
-                "timestamp": result.timestamp.isoformat(),
-            }
-            self.cache.set(cache_key, cache_data, ttl=300)
-            logger.debug(f"Cached text product explanation: {cache_key}")
-
-        return result
-
-    async def explain_afd(
-        self,
-        afd_text: str,
-        location_name: str,
-        style: ExplanationStyle = ExplanationStyle.DETAILED,
-        preserve_markdown: bool = False,
-    ) -> ExplanationResult:
-        """
-        Generate a plain-language explanation of an Area Forecast Discussion.
-
-        Thin wrapper around :meth:`explain_text_product` with
-        ``product_type="AFD"``. Kept as a public method to preserve the
-        existing call-site signature.
-        """
-        return await self.explain_text_product(
-            afd_text,
-            "AFD",
-            location_name,
-            style=style,
-            preserve_markdown=preserve_markdown,
-        )
-
-    def _call_openrouter(
-        self, system_prompt: str, user_prompt: str, model_override: str | None = None
-    ) -> dict[str, Any]:
-        """
-        Make synchronous call to OpenRouter API.
-
-        Args:
-            system_prompt: System message for the AI
-            user_prompt: User message with weather data
-            model_override: Optional model to use instead of configured model
-
-        Returns:
-            Dict with content, model, and token counts
-
-        Raises:
-            Various AIExplainerError subclasses based on error type
-
-        """
-        try:
-            client = self._get_client()
-            model = model_override or self.get_effective_model()
-
-            logger.info(
-                f"OpenRouter request: model={model}, system_prompt_len={len(system_prompt)}, user_prompt_len={len(user_prompt)}"
-            )
-
-            # Build extra_body with fallback models for free tier
-            # Only use fallbacks for default model, not user-configured models
-            extra_body = {}
-            if model == DEFAULT_FREE_MODEL and ":free" in model:
-                # Use OpenRouter's native models parameter for automatic fallback
-                fallback_models = get_available_free_models(exclude_model=model)
-                if fallback_models:
-                    extra_body["models"] = fallback_models
-                    logger.debug(f"Using fallback models: {fallback_models}")
-
-            response = client.chat.completions.create(
-                model=model,
-                messages=[
-                    {"role": "system", "content": system_prompt},
-                    {"role": "user", "content": user_prompt},
-                ],
-                max_tokens=4000,  # Increased for models with thinking/reasoning features
-                extra_headers={
-                    "HTTP-Referer": "https://accessiweather.orinks.net",
-                    "X-Title": "AccessiWeather",
-                },
-                extra_body=extra_body if extra_body else None,
-            )
-
-            # Log full response for debugging
-            logger.debug(f"OpenRouter raw response: {response}")
-
-            # Extract content - handle potential None values
-            if not response.choices:
-                logger.warning(
-                    f"OpenRouter returned empty choices. Full response: model={response.model}, id={getattr(response, 'id', 'N/A')}, usage={response.usage}"
-                )
-                return {
-                    "content": "",
-                    "model": response.model or "unknown",
-                    "total_tokens": 0,
-                    "prompt_tokens": 0,
-                    "completion_tokens": 0,
-                }
-
-            content = response.choices[0].message.content
-            finish_reason = getattr(response.choices[0], "finish_reason", "unknown")
-
-            if content is None:
-                logger.warning(
-                    f"OpenRouter returned None content. finish_reason={finish_reason}, model={response.model}"
-                )
-                content = ""
-            elif len(content.strip()) < 20:
-                logger.warning(
-                    f"OpenRouter returned short content ({len(content)} chars). finish_reason={finish_reason}, content={content[:100]!r}"
-                )
-
-            logger.info(
-                f"OpenRouter response: model={response.model}, content_len={len(content)}, finish_reason={finish_reason}"
-            )
-
-            # Handle potential None usage
-            usage = response.usage
-            return {
-                "content": content,
-                "model": response.model or "unknown",
-                "total_tokens": usage.total_tokens if usage else 0,
-                "prompt_tokens": usage.prompt_tokens if usage else 0,
-                "completion_tokens": usage.completion_tokens if usage else 0,
-            }
-
-        except Exception as e:
-            error_message = str(e).lower()
-            original_error = str(e)
-
-            # Map specific errors to custom exceptions
-            # Note: Check API key errors FIRST to avoid false matches on "connection" in suggestion text
-
-            # Missing API key (check first - most common user error)
-            if "api key required" in error_message or "api key" in error_message:
-                raise InvalidAPIKeyError(
-                    "OpenRouter API key is required.\n\n"
-                    "Please add your API key in Settings → AI Explanations.\n"
-                    "Get a free key at: openrouter.ai/keys"
-                ) from e
-
-            # Invalid API key / authentication errors (401)
-            if "invalid api key" in error_message or "authentication" in error_message:
-                raise InvalidAPIKeyError(
-                    "Your OpenRouter API key is invalid.\n\n"
-                    "Please check Settings → AI Explanations and verify your API key.\n"
-                    "Get a free key at: openrouter.ai/keys"
-                ) from e
-
-            if "401" in error_message or "unauthorized" in error_message:
-                raise InvalidAPIKeyError(
-                    "API key authentication failed.\n\n"
-                    "Your API key may be expired or incorrectly entered.\n"
-                    "Please check Settings → AI Explanations."
-                ) from e
-
-            # Insufficient credits
-            if "insufficient" in error_message or "no credits" in error_message:
-                raise InsufficientCreditsError(
-                    "Your OpenRouter account has no funds.\n\n"
-                    "Options:\n"
-                    "• Add credits at openrouter.ai/credits\n"
-                    "• Switch to a free model in Settings → AI Explanations"
-                ) from e
-
-            # Rate limiting (429) - check for status code AND common phrases
-            if (
-                "429" in error_message
-                or "rate limit" in error_message
-                or "too many requests" in error_message
-                or "rate-limited" in error_message
-            ):
-                model_used = model_override or self.get_effective_model()
-                is_free = ":free" in model_used
-                raise RateLimitError(
-                    "Rate limit exceeded.\n\n"
-                    + (
-                        "Free models share rate limits with all users and may be busy.\n\n"
-                        "Options:\n"
-                        "• Wait a few minutes and try again\n"
-                        "• Add credits to get your own rate limit\n"
-                        "• Switch to a paid model for faster access"
-                        if is_free
-                        else "Please wait a few minutes and try again."
-                    )
-                ) from e
-
-            # Timeout errors - check before generic network errors
-            if "timed out" in error_message or "timeout" in error_message:
-                raise RequestTimeoutError(
-                    "Request timed out.\n\n"
-                    "The AI service is taking too long to respond.\n"
-                    "This usually means the servers are busy. Please try again."
-                ) from e
-
-            # Network/connection errors - use specific phrases to avoid false matches
-            if (
-                "502" in error_message
-                or "503" in error_message
-                or "network error" in error_message
-                or "connection refused" in error_message
-                or "connection reset" in error_message
-            ):
-                raise NetworkError(
-                    "Network connection error.\n\n"
-                    "Could not reach the AI service. This is usually temporary.\n"
-                    "Please check your internet connection and try again."
-                ) from e
-
-            # Model not found (404)
-            if (
-                "404" in error_message
-                or "not found" in error_message
-                or "no endpoints found" in error_message
-                or "does not exist" in error_message
-            ):
-                model_used = model_override or self.get_effective_model()
-                raise InvalidModelError(
-                    f"The AI model '{model_used}' was not found.\n\n"
-                    "It may have been removed or renamed by OpenRouter.\n"
-                    "Please go to Settings → AI Explanations and select a different model."
-                ) from e
-
-            # Generic error - log full details and show user-friendly message
-            logger.error(f"OpenRouter API error: {e}", exc_info=True)
-            raise AIExplainerError(
-                f"Unable to generate explanation.\n\n"
-                f"Error: {original_error}\n\n"
-                "If this persists, try:\n"
-                "• Checking your internet connection\n"
-                "• Selecting a different AI model in Settings\n"
-                "• Trying again in a few minutes"
-            ) from e
-
-    def _estimate_cost(self, model: str, token_count: int) -> float:
-        """
-        Estimate cost based on model and token count.
-
-        Args:
-            model: Model identifier
-            token_count: Total tokens used
-
-        Returns:
-            Estimated cost in USD (0.0 for free models)
-
-        """
-        # Free models have no cost
-        if ":free" in model:
-            return 0.0
-
-        # Rough estimates for common models (per 1M tokens)
-        # These are approximations - actual costs vary
-        cost_per_million = {
-            "openrouter/auto": 0.5,  # Average estimate
-            "gpt-4": 30.0,
-            "gpt-3.5-turbo": 0.5,
-            "claude-3-opus": 15.0,
-            "claude-3-sonnet": 3.0,
-            "claude-3-haiku": 0.25,
-        }
-
-        # Find matching cost or use default
-        rate = 0.5  # Default rate
-        for model_prefix, model_rate in cost_per_million.items():
-            if model_prefix in model:
-                rate = model_rate
-                break
-
-        return (token_count / 1_000_000) * rate
-
-    async def validate_api_key(self, api_key: str) -> bool:
-        """
-        Test if API key is valid by making a minimal API call.
-
-        Args:
-            api_key: API key to validate
-
-        Returns:
-            True if valid, False otherwise
-
-        """
-        import asyncio
-
-        # Temporarily set the API key
-        original_key = self.api_key
-        self.api_key = api_key
-        self._client = None  # Reset client to use new key
-
-        try:
-            # Make a minimal API call
-            await asyncio.to_thread(
-                self._call_openrouter,
-                "You are a test assistant.",
-                "Say 'OK' if you can hear me.",
-            )
-            return True
-        except InvalidAPIKeyError:
-            return False
-        except Exception as e:
-            logger.warning(f"API key validation failed: {e}")
-            return False
-        finally:
-            # Restore original key
-            self.api_key = original_key
-            self._client = None
-
-    @staticmethod
-    async def validate_model_id(model_id: str) -> bool:
-        """
-        Validate that a model ID exists in OpenRouter's model list.
-
-        Args:
-            model_id: The model ID to validate
-
-        Returns:
-            True if the model exists, False otherwise
-
-        """
-        from .api.openrouter_models import OpenRouterModelsClient
-
-        client = OpenRouterModelsClient()
-        return await client.validate_model_id(model_id)
-
-    @staticmethod
-    async def get_valid_free_models() -> list[str]:
-        """
-        Get a list of valid free model IDs from OpenRouter.
-
-        Returns:
-            List of free model IDs
-
-        """
-        from .api.openrouter_models import OpenRouterModelsClient
-
-        client = OpenRouterModelsClient()
-        models = await client.get_free_models()
-        return [m.id for m in models]
-
-    @staticmethod
-    async def validate_and_get_fallback(model_id: str) -> tuple[str, bool]:
-        """
-        Validate a model ID and return a fallback if invalid.
-
-        Args:
-            model_id: The model ID to validate
-
-        Returns:
-            Tuple of (valid_model_id, was_fallback_used)
-            If the model is valid, returns (model_id, False)
-            If invalid, returns (DEFAULT_FREE_MODEL, True)
-
-        """
-        # Special cases that are always valid
-        if model_id in ("auto", "openrouter/auto", DEFAULT_FREE_MODEL):
-            return model_id, False
-
-        from .api.openrouter_models import OpenRouterModelsClient
-
-        client = OpenRouterModelsClient()
-        is_valid = await client.validate_model_id(model_id)
-
-        if is_valid:
-            return model_id, False
-        logger.warning(
-            f"Model '{model_id}' not found in OpenRouter. "
-            f"Falling back to default: {DEFAULT_FREE_MODEL}"
-        )
-        return DEFAULT_FREE_MODEL, True
-
-    def select_best_data_source(self, sources: list[dict[str, Any]]) -> dict[str, Any] | None:
-        """
-        Select the most recent data source from multiple sources.
-
-        Args:
-            sources: List of weather data sources with optional timestamps
-
-        Returns:
-            The source with the most recent timestamp, or None if empty
-
-        """
-        if not sources:
-            return None
-
-        # Filter sources with valid timestamps
-        sources_with_timestamps = [s for s in sources if s.get("timestamp") is not None]
-
-        if not sources_with_timestamps:
-            # If no timestamps, return first source
-            return sources[0] if sources else None
-
-        # Sort by timestamp descending and return most recent
-        sorted_sources = sorted(
-            sources_with_timestamps,
-            key=lambda s: s["timestamp"],
-            reverse=True,
-        )
-
-        return sorted_sources[0]
 
 
 def has_valid_api_key(api_key: str | None) -> bool:
